Fix problem when syncing with a delta snapshot
The problem was that we are setting the decided index before syncing with
the snapshot. Due to this, create_decided_snapshot creates an empty Snapshot
because the range 0..15 would be out of bounds. By setting the decided index
in the internal storage after creating the snapshot fixes the problem.

Moreover, this commit fixes a problem where the PartialEq of the ValueSnapshot
was ignoring the snapshotted field. Therefore, the above problem wasn't visible
in the tests.