Change Storage contract to fail when reading missing journal entries
Before the contract was to return `None` if we try to read a non-existing
journal entry. This is problematic because it might hide programming bugs
as `None` will be treated as a valid value in the upper layers. Now, the
Storage contract is that it should return an error if one tries to access
a non-existant log entry.