Let OmniPaxos::read_entries return Vec<LogEntry>
Before OmniPaxos::read_entries returned an Option<Vec<LogEntry>>. This was
problematic because the method returned None if there was an underlying storage
error or if the range was empty. This ambiguity was removed by letting this
method only return a Vec<LogEntry>. If a storage error occurs, then this method
will now panic.